+++ Core/Changelog/7.5.3.lua
@@ -2,22 +2,9 @@ local TXUI, F, E, I, V, P, G = unpack((select(2, ...)))
 
 TXUI.Changelog["7.5.3"] = {
   HOTFIX = true,
-  RELEASE_DATE = "",
+  RELEASE_DATE = "Aug 21, 2026",
   CHANGES = {
-    "* Breaking changes",
-
-    "* New features",
-
-    "* Enhancements",
-
     "* Bug fixes",
-
-    "* Profile updates",
-
-    "* Documentation",
-
-    "* Settings refactoring",
-
-    "* Development improvements",
+    F.String.Retail() .. "More secret errors -.-",
   },
 }