+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-# ToxiUI v7.5.3 - 2026-08-19
+# ToxiUI v7.5.3 - 2026-08-21
 
 **Full Changelog**: <https://github.com/Toxicom/toxiui/compare/v7.5.2...v7.5.3>
+
+## 🐛 Bug Fixes
+
+- [Retail] More secret errors -.-